// frontend/src/client/types.gen.ts
export type BankAccountPublic = {
    bank_name: string;
    account_name: string;
    account_number: string;
    account_type: string;
    opening_balance?: number;
    opening_date?: string;
    notes?: string;
    company_id?: (string | null);
    user_id?: (string | null);
    id: string;
    created_at?: (string | null);
};
// ...
export type DocumentPublic = {
    company_id?: (string | null);
    user_id?: (string | null);
    docno: string;
    doctype: string;
    client_id?: (string | null);
    date?: string;
    title: string;
    item?: (Array<{
    [key: string]: unknown;
}> | {
    [key: string]: unknown;
});
    price_calculation?: {
        [key: string]: unknown;
    };
    remark?: string;
    status: string;
    validity?: (string | null);
    duedate?: (string | null);
    generated_from_recurring_id?: (string | null);
    stock_deducted?: boolean;
    id: string;
};
// ...
export type ProductPublic = {
    product_name: string;
    description?: (string | null);
    sku?: (string | null);
    category?: string;
    cost_price?: number;
    sell_price?: number;
    supplier?: (string | null);
    stock_quantity?: number;
    unit_type?: string;
    company_id?: (string | null);
    user_id?: (string | null);
    id: string;
    created_at?: (string | null);
};
// ...
export type PurchasePublic = {
    date: string;
    due_date?: (string | null);
    supplier_name: string;
    category: string;
    company_id?: (string | null);
    user_id?: (string | null);
    amount: number;
    tax: number;
    final_amount: number;
    status: string;
    invoice_no?: (string | null);
    notes?: string;
    id: string;
};
// ...
export type SalePublic = {
    date?: string;
    channel: string;
    notes?: string;
    company_id?: (string | null);
    user_id?: (string | null);
    gross_amount: number;
    discount: number;
    net_sales: number;
    cancel_amount: number;
    short_over: number;
    refund: number;
    final_amount: number;
    status: string;
    id: string;
};